requirements.txt
python-dotenv
minio
pytelegrambotapi
Pillow
pytest
isort
flake8